requirements.txt

Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
gunicorn==23.0.0
